Add search box for searching local/remote branches or tags.
This commit adds a new search box into left repository panel.
This text box allows searching for local branches, remote branches or tags simultaniously with same text. It has a regex support.
Search is performed on timer to expect user complete typing and not spamming the search for each input symbol. While typing the regex pattern the search can be triggered with invalid pattern, in this case the search results are reset and no exception is propagaded.