enhance: depend on SO in rpm package  (#688)
* enhance: ensure LF line endings used with package manifests
* enhance: depend on shared object instead of package

Back in 211c263, I changed the dependencies to depend on package instead of shared object. This worked well, but introduced some problems for distros other than Fedora (81f76f0). I did this because of problems with the arm64 package, but now I think `__isa_bits` is a reasonable workaround.

* refactor: remove pointless variable checks

`set -u` was introduced in 6461765