fix: searching for tree/blob hash never-ending (#1472)
Searching for a tree or blob hash instead of a commit hash created a
seemingly never-ending search (internally an exception was thrown).

Examples for non-ending searches in the SourceGit repository itself are
searching for 6b83a0a (that is the tree "src" of commit 190d2ee) or
e686a53 (that is the file ".gitignore" in the tree of commit 190d2ee).