fix: handling of last line without newline in hunks
Motivation: Some editors (depending on editor configuration) change the
last line of a file by adding a missing newline.  But in order to keep
diffs at a minimum, it's sometimes advisable to not include this change
into the commit.  This should be possible by discarding the hunk with
this change.

Now it is possible to discard (or stage/unstage) such a change.
Previously discarding such a change was ignored, when there were other
changes in the same file.