feat: prune on demand - with respect to git config as default
the project already have global setting that is fully synced with git config.
but sometimes there is a need to prune on demand.
in these scenarios there is no way to do it via SourceGit and I find myself everytime switch to terminal
I also saw there was another PR suggesting a fix for that - but this PR respects the repo config and git config